+++ src/feedmark/parser.py
@@ -2,8 +2,10 @@
 # encoding: UTF-8
 
 from datetime import datetime
+from collections import OrderedDict
 import re
 
+from feedmark.formats.markdown import markdown_to_html5, markdown_to_html5_deep
 from feedmark.utils import quote
 
 
@@ -31,7 +33,7 @@ def rewrite_reference_links(refdex, reference_links):
 class Document(object):
     def __init__(self, title):
         self.title = title
-        self.properties = {}
+        self.properties = OrderedDict()
 
         self.preamble = None
         self.sections = []
@@ -44,12 +46,46 @@ class Document(object):
         for section in self.sections:
             section.reference_links = rewrite_reference_links(refdex, section.reference_links)
 
+    def global_reference_links(self):
+        reference_links = []
+        reference_links.extend(self.reference_links)
+        for section in self.sections:
+            reference_links.extend(section.reference_links)
+        return reference_links
+
+    def to_json_data(self, **kwargs):
+
+        if kwargs.get('htmlize', False):
+            if 'reference_links' not in kwargs:
+                kwargs['reference_links'] = self.global_reference_links()
+            preamble = markdown_to_html5(self.preamble, reference_links=kwargs['reference_links'])
+            properties = markdown_to_html5_deep(self.properties, reference_links=kwargs['reference_links'])
+        else:
+            preamble = self.preamble
+            properties = self.properties
+
+        if kwargs.get('ordered', False):
+            properties_list = []
+            for key, value in properties.items():
+                properties_list.append([key, value])
+            properties = properties_list
+        else:
+            properties = dict(properties)
+
+        return {
+            'filename': self.filename,
+            'title': self.title,
+            'properties': properties,
+            'preamble': preamble,
+            'sections': [s.to_json_data(**kwargs) for s in self.sections],
+        }
+
 
 class Section(object):
     def __init__(self, title):
         self.document = None
         self.title = title
-        self.properties = {}
+        self.properties = OrderedDict()
 
         self.lines = []
 
@@ -58,12 +94,6 @@ class Section(object):
         if self.document:
             s += " of " + str(self.document)
         return s
-
-    def add_line(self, line):
-        self.lines.append(line.rstrip())
-
-    def set(self, key, value):
-        self.properties[key] = value
 
     @property
     def body(self):
@@ -89,6 +119,31 @@ class Section(object):
         from markdown.extensions.toc import slugify
 
         return slugify(self.title, '-')
+
+    def to_json_data(self, **kwargs):
+
+        if kwargs.get('htmlize', False):
+            body = markdown_to_html5(self.body, reference_links=kwargs['reference_links'])
+            properties = markdown_to_html5_deep(self.properties, reference_links=kwargs['reference_links'])
+        else:
+            body = self.body
+            properties = self.properties
+
+        if kwargs.get('ordered', False):
+            properties_list = []
+            for key, value in properties.items():
+                properties_list.append([key, value])
+            properties = properties_list
+        else:
+            properties = dict(properties)
+
+        return {
+            'title': self.title,
+            'anchor': self.anchor,
+            'images': self.images,
+            'properties': properties,
+            'body': body,
+        }
 
 
 class Parser(object):
@@ -169,7 +224,7 @@ class Parser(object):
         raise ValueError('Expected property')
 
     def parse_properties(self):
-        properties = {}
+        properties = OrderedDict()
         while self.is_blank_line() or self.is_property_line():
             if self.is_property_line():
                 kind, key, val = self.parse_property()